requirements/requirements_eval_meddr.txt
--index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
accelerate==0.34.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.11.0
async-lru==2.0.5
async-timeout==5.0.1
asyncio-atexit==1.0.1
attrs==25.4.0
av==15.1.0
backoff==2.2.1
beautifulsoup4==4.14.2
bitsandbytes==0.45.2
black==25.9.0
Brotli==1.1.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
DataProperty==1.1.0
datasets==3.6.0
decord==0.6.0
deepspeed==0.18.0
Deprecated==1.2.18
dill==0.3.8
distlib==0.4.0
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evaluate==0.4.6
exceptiongroup==1.3.0
filelock==3.19.1
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.3/flash_attn-2.7.3+cu12torch2.6cxx11abiFALSE-cp39-cp39-linux_x86_64.whl#sha256=1ef6c7b99c12ef0fbb051e5d28f88a4092afac668e6cc15c1d89ac13eefac852
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gdown==5.2.0
gdrive==0.1.5
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-api-python-client==2.43.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.10
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
identify==2.6.15
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflate64==1.0.3
inquirerpy==0.3.4
isort==6.1.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==4.0.0
kiwisolver==1.4.7
latex2sympy2==1.9.1
lazy_loader==0.4
loguru==0.7.3
lxml==6.0.2
MarkupSafe==2.1.5
matplotlib==3.9.4
mbstrdecoder==1.1.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
multivolumefile==0.2.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
ninja==1.13.0
nltk==3.9.2
nodeenv==1.9.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
openai==2.5.0
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-instrumentation-urllib==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pathvalidate==3.3.1
peft==0.10.0
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.4.0
portalocker==3.2.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==3.20.0
psutil==5.9.8
py-cpuinfo==9.0.0
py7zr==1.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pybind11==3.0.1
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pynrrd==1.1.3
pyparsing==3.2.5
pyppmd==1.2.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
pyzstd==0.18.0
qwen-vl-utils==0.0.14
rarfile==4.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sacrebleu==2.5.1
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
SecretStorage==3.3.3
sentence-transformers==5.1.1
sentencepiece==0.1.99
sentry-sdk==2.42.0
shortuuid==1.0.13
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlitedict==2.1.0
sympy==1.13.1
synapseclient==4.10.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.1.2
tensorboardX==2.6.4
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.12.0
timm==0.9.12
tokenizers==0.15.1
tomli==2.3.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.37.2
transformers-stream-generator==0.0.5
triton==3.2.0
typepy==1.3.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
versioneer==0.29
virtualenv==20.35.3
wandb==0.22.2
wcwidth==0.2.14
wrapt==1.17.3
xxhash==3.6.0
yacs==0.1.8
yarl==1.22.0
yt-dlp==2025.10.14
zipp==3.23.0
zss==1.2.0
zstandard==0.25.0
